requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.1
langchain>=0.1.0
langchain-community>=0.0.13
langchain-huggingface>=0.0.6
faiss-cpu>=1.7.4
transformers>=4.36.2
torch>=2.1.2
huggingface-hub>=0.20.3
PyMuPDF>=1.23.8
PyPDF2>=3.0.1
python-docx>=1.0.1
pandas>=2.1.4
openpyxl>=3.1.2
xlrd>=2.0.1
python-pptx>=0.6.22
PyYAML>=6.0.1
Pillow>=10.2.0
google-cloud-vision>=3.5.0
google-cloud-storage>=2.14.0
packaging>=23.2
pydantic<2.0.0
requests
# For Gemini API
google-generativeai
# For CORS and web
flask-cors
# For image conversion (pdf2image)
poppler-utils